Move EmailFollowup in the models.common.
