+++ huntflow_api_client/models/request/applicants.py
@@ -3,7 +3,12 @@ from typing import List, Literal, Optional
 
 from pydantic import BaseModel, EmailStr, Field, PositiveInt
 
-from huntflow_api_client.models.common import Applicant, EmailRecipient, JsonRequestModel
+from huntflow_api_client.models.common import (
+    Applicant,
+    EmailFollowup,
+    EmailRecipient,
+    JsonRequestModel,
+)
 from huntflow_api_client.models.consts import (
     CalendarEventReminderMethod,
     CalendarEventType,
@@ -67,17 +72,6 @@ class CalendarEventAttendee(BaseModel):
 
     class Config:
         allow_population_by_field_name = True
-
-
-class EmailFollowup(BaseModel):
-    id: PositiveInt = Field(..., description="Followup ID")
-    account_member_template: int = Field(..., description="Email template ID")
-    html: str = Field(..., description="Email content (HTML)")
-    days: int = Field(
-        ...,
-        gte=1,
-        description="The number of days after which to send a followup if there is no response",
-    )
 
 
 class ApplicantLogEmail(BaseModel):
